Correcting boolean parser arguments, adding extra prints with debug info

+++ Duplicates Removal/remove_duplicates.py
@@ -15,8 +15,8 @@ def get_args():
     parser.add_argument('--encoder', type=str, choices=['CNN, DHash'], default='CNN', help='Type of encoding for features comparison')
     parser.add_argument('--threshold', type=float, default=0.9,
         help='For hashing: max_distance_threshold, for CNN: min_similarity_threshold (note: the logic of these differs, refer to documentation)')
-    parser.add_argument('--display_removed', type=bool, default='False', help='Show images that will be removed, requires tkinter')
-    parser.add_argument('--dry_run', type=bool, default=False, help='Does not remove images in dry run')
+    parser.add_argument('--display_removed', action='store_true', help='Show images that will be removed, requires tkinter')
+    parser.add_argument('--dry_run', action='store_true', help='Does not remove images in dry run')
     args = parser.parse_args()
     return args
 
@@ -34,6 +34,7 @@ def remove_duplicates(args):
     print(f'Number of duplicates to remove: {len(duplicates)}')
 
     if args.display_removed:
+        print('Displaying duplicate images')
         duplicates_map = []
         if args.encoder == 'CNN':
             duplicates_map = encoder.find_duplicates(encoding_map=encodings, min_similarity_threshold=args.threshold)
@@ -44,11 +45,14 @@ def remove_duplicates(args):
         for dup in duplicates:
             plot_duplicates(args.dir, duplicates_map, dup)
 
-    os.makedirs(f'{args.dir}/duplicates', exist_ok=True)
+    if not args.dry_run:
+        print('Moving duplicate images to subdirectory "duplicates"')
+        os.makedirs(f'{args.dir}/duplicates', exist_ok=True)
 
-    for dup in duplicates:
-        os.rename(f'{args.dir}/{dup}', f'{args.dir}/duplicates/{dup}')
+        for dup in duplicates:
+            os.rename(f'{args.dir}/{dup}', f'{args.dir}/duplicates/{dup}')
 
 
 args = get_args()
+print(args)
 remove_duplicates(args)
